INFOPLAT-2071 Handles creating emitter based on if chip ingress is configured
INFOPLAT-2071 Adjusts code comment

INFOPLAT-2071 Refactor emitters
