+++ pkg/beholder/chip_ingress_emitter.go
@@ -2,33 +2,38 @@ package beholder
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/smartcontractkit/chainlink-common/pkg/beholder/pb"
 	"github.com/smartcontractkit/chainlink-common/pkg/chipingress"
+	"google.golang.org/grpc/credentials/insecure"
 )
 
 type chipIngressEmitter struct {
 	client chipingress.ChipIngressClient
 }
 
-// dualSourceEmitter emits both to chip ingress and to the otel collector
-// this is to help transition from sending custom messages via OTLP to instead use chip-ingress
-// we want to send to both during the transition period, then cutover to using
-// chipIngressEmitter only
-type dualSourceEmitter struct {
-	chipIngressEmitter Emitter
-	otelCollectorEmitter Emitter 
-}
-
 func NewChipIngressEmitter(client chipingress.ChipIngressClient) Emitter {
 	return &chipIngressEmitter{client: client,}
 }
 
-func NewDualSourceEmitter(chipIngressClient chipingress.ChipIngressClient, otelCollectorEmitter Emitter) Emitter {
-	return &dualSourceEmitter{
-		chipIngressEmitter: NewChipIngressEmitter(chipIngressClient),
-		otelCollectorEmitter: otelCollectorEmitter,
+func NewChipIngressClient(cfg Config) (chipingress.ChipIngressClient, error) {
+
+	if cfg.ChipIngressEmitterGRPCEndpoint == "" {
+		return nil, fmt.Errorf("missing chip ingress emitter gRPC endpoint")
 	}
+
+	// TODO: add support for csa auth signing interceptor
+	// We should add csa signed headers, that will be authenticated on the server-side
+	client, err := chipingress.NewChipIngressClient(
+		cfg.ChipIngressEmitterGRPCEndpoint,
+		chipingress.WithTransportCredentials(insecure.NewCredentials()),
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	return client, nil
 }
 
 func (c *chipIngressEmitter) Emit(ctx context.Context, body []byte, attrKVs ...any) error {
@@ -46,21 +51,6 @@ func (c *chipIngressEmitter) Emit(ctx context.Context, body []byte, attrKVs ...a
 	if err != nil {
 		return err
 	}
-
-	return nil
-}
-
-func (d *dualSourceEmitter) Emit(ctx context.Context, body []byte, attrKVs ...any) error {
-	
-	// Emit via OTLP first
-	if err := d.otelCollectorEmitter.Emit(ctx, body, attrKVs...); err != nil {
-		return err
-	}
-
-	if err := d.chipIngressEmitter.Emit(ctx, body, attrKVs...); err != nil {
-		return err
-	}
-
 
 	return nil
 }
